+++ src/repo/mongoose/model/action.ts
@@ -119,6 +119,7 @@ export default mongoose.model('Action', schema)
         // tslint:disable-next-line:no-single-line-block-comment
         /* istanbul ignore next */
         if (error !== undefined) {
+            // tslint:disable-next-line:no-console
             console.error(error);
         }
     });