+++ src/repo/mongoose/model/aggregateSale.ts
@@ -135,6 +135,7 @@ export default mongoose.model('AggregateSale', schema)
         // tslint:disable-next-line:no-single-line-block-comment
         /* istanbul ignore next */
         if (error !== undefined) {
+            // tslint:disable-next-line:no-console
             console.error(error);
         }
     });