+++ src/repo/mongoose/model/film.ts
@@ -46,6 +46,7 @@ export default mongoose.model('Film', schema)
         // tslint:disable-next-line:no-single-line-block-comment
         /* istanbul ignore next */
         if (error !== undefined) {
+            // tslint:disable-next-line:no-console
             console.error(error);
         }
     });