+++ src/repo/mongoose/model/order.ts
@@ -121,6 +121,7 @@ export default mongoose.model('Order', schema)
         // tslint:disable-next-line:no-single-line-block-comment
         /* istanbul ignore next */
         if (error !== undefined) {
+            // tslint:disable-next-line:no-console
             console.error(error);
         }
     });