+++ src/repo/mongoose/model/reservation.ts
@@ -207,6 +207,7 @@ export default mongoose.model('Reservation', schema)
         // tslint:disable-next-line:no-single-line-block-comment
         /* istanbul ignore next */
         if (error !== undefined) {
+            // tslint:disable-next-line:no-console
             console.error(error);
         }
     });