+++ src/repo/mongoose/model/stock.ts
@@ -70,6 +70,7 @@ export default mongoose.model('Stock', schema)
         // tslint:disable-next-line:no-single-line-block-comment
         /* istanbul ignore next */
         if (error !== undefined) {
+            // tslint:disable-next-line:no-console
             console.error(error);
         }
     });