+++ src/repo/mongoose/model/task.ts
@@ -87,6 +87,7 @@ export default mongoose.model('Task', schema)
         // tslint:disable-next-line:no-single-line-block-comment
         /* istanbul ignore next */
         if (error !== undefined) {
+            // tslint:disable-next-line:no-console
             console.error(error);
         }
     });