+++ src/repo/mongoose/model/theater.ts
@@ -33,6 +33,7 @@ export default mongoose.model('Theater', schema)
         // tslint:disable-next-line:no-single-line-block-comment
         /* istanbul ignore next */
         if (error !== undefined) {
+            // tslint:disable-next-line:no-console
             console.error(error);
         }
     });